Update hw3.properties
Even if it's a test account, you shouldn't share username and password as clear text. I recommend deleting the hw3.properties file from the project so that the login information history cannot be traced, and reloading it as the login information will not be visible. may be you can read them from txt file exc.